+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* New: `wrap_text` for using custom control sequences.
 * Fix broken randomness in `html_id`, also using a fixed format (`[prefix]-[five digits]-[five digits]`)
 
 ## [0.3.0] - 2022-01-11